Extract toString generation helpers to new file
